backend/requirements.txt
fastapi
uvicorn
paddleocr
pillow
